sources/ricable/polyglot-devenv/devpod-automation/templates/.claude-core/mcp/requirements-mcp.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.6.5
python-multipart>=0.0.6
google-generativeai>=0.3.0
aiohttp>=3.9.0